Consumer 추가

+++ api/src/main/java/org/delivery/api/domain/token/service/TokenService.java
@@ -31,7 +31,7 @@ public class TokenService {
 
     public Long validationToken(String token) {
         Map<String, Object> map = tokenHelper.validationTokenWithThrow(token);
-        Long userId = (Long) map.get("userId");
+        Integer userId = (Integer) map.get("userId");
         Objects.requireNonNull(userId, () -> {
             throw new ApiException(ErrorCode.NULL_POINT);
         });
